requirements.txt
Package                Version
---------------------- -----------
aiogram                2.25.1
aiohttp                3.8.4
aioschedule            0.5.2
aiosignal              1.3.1
app-store-scraper      0.3.5
APScheduler            3.10.0
async-timeout          4.0.2
attrs                  22.2.0
Babel                  2.9.1
certifi                2022.12.7
chardet                3.0.4
charset-normalizer     3.0.1
colorama               0.4.6
contourpy              1.0.7
cycler                 0.11.0
fonttools              4.38.0
frozenlist             1.3.3
google-play-scraper    1.2.3
idna                   2.10
itunes-app-scraper-dmi 0.9.5
kiwisolver             1.4.4
magic-filter           1.0.9
matplotlib             3.7.0
multidict              6.0.4
numpy                  1.24.2
packaging              23.0
pandas                 1.5.3
Pillow                 9.4.0
pip                    22.3.1
plotly                 5.13.0
psutil                 5.9.4
pyparsing              3.0.9
python-dateutil        2.8.2
pytz                   2022.7.1
pytz-deprecation-shim  0.1.0.post0
requests               2.23.0
setuptools             65.5.0
six                    1.16.0
tenacity               8.2.1
tqdm                   4.64.1
tzdata                 2022.7
tzlocal                4.2
urllib3                1.25.11
yarl                   1.8.2
